+++ src/main/java/uk/co/techblue/alfresco/dto/user/AuthorityQuery.java
@@ -17,14 +17,12 @@ package uk.co.techblue.alfresco.dto.user;
 
 import javax.ws.rs.QueryParam;
 
-import uk.co.techblue.alfresco.dto.BaseDto;
 import uk.co.techblue.alfresco.dto.common.AuthorityType;
-import uk.co.techblue.alfresco.dto.common.GroupSorter;
 
 /**
  * The Class AuthorityQuery.
  */
-public class AuthorityQuery extends BaseDto {
+public class AuthorityQuery extends BaseQuery {
 
     /** The Constant serialVersionUID. */
     private static final long serialVersionUID = -3979600916433012399L;
@@ -32,18 +30,6 @@ public class AuthorityQuery extends BaseDto {
     /** The authority type. */
     @QueryParam("authorityType")
     private AuthorityType authorityType;
-
-    /** The max items. */
-    @QueryParam("maxItems")
-    private Integer maxItems;
-
-    /** The skip count. */
-    @QueryParam("skipCount")
-    private Integer skipCount;
-
-    /** The sort by. */
-    @QueryParam("sortBy")
-    private GroupSorter sortBy;
 
     /**
      * Gets the authority type.
@@ -61,60 +47,6 @@ public class AuthorityQuery extends BaseDto {
      */
     public void setAuthorityType(final AuthorityType authorityType) {
         this.authorityType = authorityType;
-    }
-
-    /**
-     * Gets the max items.
-     * 
-     * @return the max items
-     */
-    public Integer getMaxItems() {
-        return maxItems;
-    }
-
-    /**
-     * Sets the max items.
-     * 
-     * @param maxItems the new max items
-     */
-    public void setMaxItems(final Integer maxItems) {
-        this.maxItems = maxItems;
-    }
-
-    /**
-     * Gets the skip count.
-     * 
-     * @return the skip count
-     */
-    public Integer getSkipCount() {
-        return skipCount;
-    }
-
-    /**
-     * Sets the skip count.
-     * 
-     * @param skipCount the new skip count
-     */
-    public void setSkipCount(final Integer skipCount) {
-        this.skipCount = skipCount;
-    }
-
-    /**
-     * Gets the sort by.
-     * 
-     * @return the sort by
-     */
-    public GroupSorter getSortBy() {
-        return sortBy;
-    }
-
-    /**
-     * Sets the sort by.
-     * 
-     * @param sortBy the new sort by
-     */
-    public void setSortBy(final GroupSorter sortBy) {
-        this.sortBy = sortBy;
     }
 
 }
