+++ src/main/java/uk/co/techblue/alfresco/dto/user/GroupQuery.java
@@ -17,13 +17,10 @@ package uk.co.techblue.alfresco.dto.user;
 
 import javax.ws.rs.QueryParam;
 
-import uk.co.techblue.alfresco.dto.BaseDto;
-import uk.co.techblue.alfresco.dto.common.GroupSorter;
-
 /**
  * The Class GroupQuery.
  */
-public class GroupQuery extends BaseDto {
+public class GroupQuery extends BaseQuery {
 
     /** The Constant serialVersionUID. */
     private static final long serialVersionUID = 4809842369740131482L;
@@ -35,18 +32,6 @@ public class GroupQuery extends BaseDto {
     /** The zone. */
     @QueryParam("zone")
     private Boolean zone;
-
-    /** The max items. */
-    @QueryParam("maxItems")
-    private Integer maxItems;
-
-    /** The skip count. */
-    @QueryParam("skipCount")
-    private Integer skipCount;
-
-    /** The sort by. */
-    @QueryParam("sortBy")
-    private GroupSorter sortBy;
 
     /**
      * Gets the short name filter.
@@ -82,60 +67,6 @@ public class GroupQuery extends BaseDto {
      */
     public void setZone(final Boolean zone) {
         this.zone = zone;
-    }
-
-    /**
-     * Gets the max items.
-     * 
-     * @return the max items
-     */
-    public Integer getMaxItems() {
-        return maxItems;
-    }
-
-    /**
-     * Sets the max items.
-     * 
-     * @param maxItems the new max items
-     */
-    public void setMaxItems(final Integer maxItems) {
-        this.maxItems = maxItems;
-    }
-
-    /**
-     * Gets the skip count.
-     * 
-     * @return the skip count
-     */
-    public Integer getSkipCount() {
-        return skipCount;
-    }
-
-    /**
-     * Sets the skip count.
-     * 
-     * @param skipCount the new skip count
-     */
-    public void setSkipCount(final Integer skipCount) {
-        this.skipCount = skipCount;
-    }
-
-    /**
-     * Gets the sort by.
-     * 
-     * @return the sort by
-     */
-    public GroupSorter getSortBy() {
-        return sortBy;
-    }
-
-    /**
-     * Sets the sort by.
-     * 
-     * @param sortBy the new sort by
-     */
-    public void setSortBy(final GroupSorter sortBy) {
-        this.sortBy = sortBy;
     }
 
 }
